Move isSubDirectory function to fileProvider class.
